Move Koalas specific methods to koalas accessor
