+++ databricks/koalas/datetimes.py
@@ -173,7 +173,7 @@ class DatetimeMethods(object):
         def pandas_dayofweek(s) -> "ks.Series[np.int64]":
             return s.dt.dayofweek
 
-        return self._data.transform_batch(pandas_dayofweek)
+        return self._data.koalas.transform_batch(pandas_dayofweek)
 
     @property
     def weekday(self) -> "ks.Series":
@@ -190,7 +190,7 @@ class DatetimeMethods(object):
         def pandas_dayofyear(s) -> "ks.Series[np.int64]":
             return s.dt.dayofyear
 
-        return self._data.transform_batch(pandas_dayofyear)
+        return self._data.koalas.transform_batch(pandas_dayofyear)
 
     @property
     def quarter(self) -> "ks.Series":
@@ -201,7 +201,7 @@ class DatetimeMethods(object):
         def pandas_quarter(s) -> "ks.Series[np.int64]":
             return s.dt.quarter
 
-        return self._data.transform_batch(pandas_quarter)
+        return self._data.koalas.transform_batch(pandas_quarter)
 
     @property
     def is_month_start(self) -> "ks.Series":
@@ -240,7 +240,7 @@ class DatetimeMethods(object):
         def pandas_is_month_start(s) -> "ks.Series[bool]":
             return s.dt.is_month_start
 
-        return self._data.transform_batch(pandas_is_month_start)
+        return self._data.koalas.transform_batch(pandas_is_month_start)
 
     @property
     def is_month_end(self) -> "ks.Series":
@@ -279,7 +279,7 @@ class DatetimeMethods(object):
         def pandas_is_month_end(s) -> "ks.Series[bool]":
             return s.dt.is_month_end
 
-        return self._data.transform_batch(pandas_is_month_end)
+        return self._data.koalas.transform_batch(pandas_is_month_end)
 
     @property
     def is_quarter_start(self) -> "ks.Series":
@@ -329,7 +329,7 @@ class DatetimeMethods(object):
         def pandas_is_quarter_start(s) -> "ks.Series[bool]":
             return s.dt.is_quarter_start
 
-        return self._data.transform_batch(pandas_is_quarter_start)
+        return self._data.koalas.transform_batch(pandas_is_quarter_start)
 
     @property
     def is_quarter_end(self) -> "ks.Series":
@@ -379,7 +379,7 @@ class DatetimeMethods(object):
         def pandas_is_quarter_end(s) -> "ks.Series[bool]":
             return s.dt.is_quarter_end
 
-        return self._data.transform_batch(pandas_is_quarter_end)
+        return self._data.koalas.transform_batch(pandas_is_quarter_end)
 
     @property
     def is_year_start(self) -> "ks.Series":
@@ -418,7 +418,7 @@ class DatetimeMethods(object):
         def pandas_is_year_start(s) -> "ks.Series[bool]":
             return s.dt.is_year_start
 
-        return self._data.transform_batch(pandas_is_year_start)
+        return self._data.koalas.transform_batch(pandas_is_year_start)
 
     @property
     def is_year_end(self) -> "ks.Series":
@@ -457,7 +457,7 @@ class DatetimeMethods(object):
         def pandas_is_year_end(s) -> "ks.Series[bool]":
             return s.dt.is_year_end
 
-        return self._data.transform_batch(pandas_is_year_end)
+        return self._data.koalas.transform_batch(pandas_is_year_end)
 
     @property
     def is_leap_year(self) -> "ks.Series":
@@ -496,7 +496,7 @@ class DatetimeMethods(object):
         def pandas_is_leap_year(s) -> "ks.Series[bool]":
             return s.dt.is_leap_year
 
-        return self._data.transform_batch(pandas_is_leap_year)
+        return self._data.koalas.transform_batch(pandas_is_leap_year)
 
     @property
     def daysinmonth(self) -> "ks.Series":
@@ -507,7 +507,7 @@ class DatetimeMethods(object):
         def pandas_daysinmonth(s) -> "ks.Series[np.int64]":
             return s.dt.daysinmonth
 
-        return self._data.transform_batch(pandas_daysinmonth)
+        return self._data.koalas.transform_batch(pandas_daysinmonth)
 
     @property
     def days_in_month(self) -> "ks.Series":
@@ -567,7 +567,7 @@ class DatetimeMethods(object):
         def pandas_normalize(s) -> "ks.Series[np.datetime64]":
             return s.dt.normalize()
 
-        return self._data.transform_batch(pandas_normalize)
+        return self._data.koalas.transform_batch(pandas_normalize)
 
     def strftime(self, date_format) -> "ks.Series":
         """
@@ -615,7 +615,7 @@ class DatetimeMethods(object):
         def pandas_strftime(s) -> "ks.Series[str]":
             return s.dt.strftime(date_format)
 
-        return self._data.transform_batch(pandas_strftime)
+        return self._data.koalas.transform_batch(pandas_strftime)
 
     def round(self, freq, *args, **kwargs) -> "ks.Series":
         """
@@ -670,7 +670,7 @@ class DatetimeMethods(object):
         def pandas_round(s) -> "ks.Series[np.datetime64]":
             return s.dt.round(freq, *args, **kwargs)
 
-        return self._data.transform_batch(pandas_round)
+        return self._data.koalas.transform_batch(pandas_round)
 
     def floor(self, freq, *args, **kwargs) -> "ks.Series":
         """
@@ -725,7 +725,7 @@ class DatetimeMethods(object):
         def pandas_floor(s) -> "ks.Series[np.datetime64]":
             return s.dt.floor(freq, *args, **kwargs)
 
-        return self._data.transform_batch(pandas_floor)
+        return self._data.koalas.transform_batch(pandas_floor)
 
     def ceil(self, freq, *args, **kwargs) -> "ks.Series":
         """
@@ -780,7 +780,7 @@ class DatetimeMethods(object):
         def pandas_ceil(s) -> "ks.Series[np.datetime64]":
             return s.dt.ceil(freq, *args, **kwargs)
 
-        return self._data.transform_batch(pandas_ceil)
+        return self._data.koalas.transform_batch(pandas_ceil)
 
     def month_name(self, locale=None) -> "ks.Series":
         """
@@ -816,7 +816,7 @@ class DatetimeMethods(object):
         def pandas_month_name(s) -> "ks.Series[str]":
             return s.dt.month_name(locale=locale)
 
-        return self._data.transform_batch(pandas_month_name)
+        return self._data.koalas.transform_batch(pandas_month_name)
 
     def day_name(self, locale=None) -> "ks.Series":
         """
@@ -852,4 +852,4 @@ class DatetimeMethods(object):
         def pandas_day_name(s) -> "ks.Series[str]":
             return s.dt.day_name(locale=locale)
 
-        return self._data.transform_batch(pandas_day_name)
+        return self._data.koalas.transform_batch(pandas_day_name)
